refactor: split QuanticsTransform.jl into focused files
